requirements.txt
edge_tts == 7.2.7
openai == 2.24.0
numpy==1.26.4
opuslib_next==1.1.5
pydub==0.25.1
funasr == 1.2.6
torch == 2.2.2
torchaudio == 2.2.2
librosa~=0.11.0
pyyaml~=6.0.3